backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
motor
dnspython
pyjwt
passlib[bcrypt]
bcrypt
